rework and create some function, generate comment from card.js
